[profiles] Fix valuedescriptor Type initialization
The original Java device SDK set the Type member of
a ValueDescriptor to the first char of the ValueProperty
Type member when creating a new ValueDescriptor. This
doesn't work with the new expanded type values, as there'd
be no way to distinguish between an int8 and int16, or
float32 and float64. So the full Type from the ValueProperty
is now used to initialize the new ValueDescriptor's Type.

Signed-off-by: Tony Espy <espy@canonical.com>