Re-factor ProtocolDriver HandleOperation function
This change re-factors the logic for handling the command
endpoint. In particular, instead of looping through the
ResourceOperations of a particular call to the command endpoint,
it now passes them all in a single call to the ProtocolDriver
and waits for all the results to be returned as a single slice.
This change also renames HandleOperation to HandleCommands.

Signed-off-by: Tony Espy <espy@canonical.com>