Accept JSON map for command PUT request params
The current implementation only accepts a JSON array
of maps for the command parms pass as the request body.
As this is wrong (see issue #141), this change modifies
the code to accept a single map of parameters, but will
also continue to accept an array of maps in order to
not break existing clients that pass an array.

Signed-off-by: Cloud Tsai <cloudxxx8@gmail.com>