bootstrap: fix OS signal handling
Applications using bootstrap fail to respond to SIGINT (ctrl-C)
because the Service Start() method is a blocking call, which prevents
the signal handling logic from running. Instead call the Start method
in a new goroutine.
Also, add a check of logging client existence in driver.Stop(),
because the logging client might be nil if SIGINT (ctrl-C) received
during clients initialization.

Fixes: https://github.com/edgexfoundry/device-sdk-go/issues/242

Signed-off-by: Cloud Tsai <cloudxxx8@gmail.com>