refactor(sdk): Refactor to use New func pattern instead of helper functions
The continued growth of this SDK needs to be kept organized and succinct. Previously functions
were exposed directly as part of the SDK which overtime became a bit larger than preferred. This
commit implements the New() pattern to create the specific instance of the transform with the
properties required and allows a developer to specify the function desired on the specific set of
transforms.

BREAKING CHANGE: Pipeline functions used in the SetPipeline() now need to be created with the provided New…() functions.

Closes #145

Signed-off-by: lenny <leonard.goodell@intel.com>