feat(configuration) : Add UseTargetTypeOfByteArray to Pipeline confguration
With the addition of the UseTargetTypeOfByteArray boolean setting the functions pipeline can be configured to accept a byte array as input to the first function in the pipeline. This is instead of the default EdgeX Event object. Having this setting in the Writable Pipeline section allows it to be changed appropriately to match the first function in the pipeline as it is changed.

closes #177

Signed-off-by: Leonard Goodell <leonard.goodell@intel.com>