Move Creation of ValueDescriptors to Metadata
The plan it to move whole ValueDescriptor management logic to Core Metadata in Geneva.  For now, there is a config called EnableValueDescriptorManagement in Core Metadata to control whether the management responsibility is under Core Metadata or Device Service SDK.
This commit is adding the logic to retrieve the is config value from
Core Metadata and skip the ValueDescriptor creation if it is true.

Fix https://github.com/edgexfoundry/device-sdk-go/issues/318

Signed-off-by: Cloud Tsai <cloudxxx8@gmail.com>