fix: Create a buffer to handle the AsyncValues
Since the DS sends the AsyncValues via the channel, so that AsyncValues needs to wait in the queue. If the reading amount getting large, it needs more time to transport those data.
To enhance performance, refactor function to create a buffer for receiving AsyncValues and using goroutine to process each AsyncValues concurrently.

Signed-off-by: weichou <weichou1229@gmail.com>