fix: update CommandExists to only check Device Commands
SDK originally check CoreCommand first and return true if given command
exists, which might lead to some potential bug because Read/WriteCmd
functions are designed for DeviceCommand which rely on ResourceOperation
to access DeviceResource.
Update the CommandExists function to only check DeviceCommand to prevent
bugs addressed in #560 and rename the ResourceOperation map structure in
profile cache to be more specific.

Signed-off-by: Chris Hung <chris@iotechsys.com>