refactor: upgrade provision package to use v2 models/clients
while loading devices defined in local configuration file, instead of
checking profile existed in cache, we now check if the profile exists
in metadata because v2 SDK only supports profile callback on update.

The profiles will eventually be added into cache when a device using it
is added via callback.

Signed-off-by: Chris Hung <chris@iotechsys.com>