feat(sdk): Custom Trigger Multi-Pipeline Support (#941)
Add service.processMessageOnRuntime to use default pipeline if configured or attempt to use topic matching logic.  Deprecate TriggerMessageProcessor and
TriggerContextBuilder and inline the existing functions on TriggerConfig as the approach used there will not work with multiple pipelines. Replace with
MessageProcessor that only takes the message envelope and builds context(s) as needed.

Signed-off-by: Alex Ullrich <alex.ullrich@gmail.com>